+++ h3ronpy/python/h3ronpy/polars/raster.py
@@ -1,9 +1,9 @@
-from ..arrow import raster as arrow_raster
+from h3ronpy import raster
 
-nearest_h3_resolution = arrow_raster.nearest_h3_resolution
-rasterize_cells = arrow_raster.rasterize_cells
+nearest_h3_resolution = raster.nearest_h3_resolution
+rasterize_cells = raster.rasterize_cells
 
-__doc__ = arrow_raster.__doc__
+__doc__ = raster.__doc__
 
 __all__ = [
     nearest_h3_resolution.__name__,
