+++ h3ronpy/python/h3ronpy/raster.py
@@ -33,11 +33,10 @@ import typing
 import numpy as np
 import pyarrow as pa
 
+from h3ronpy import DEFAULT_CELL_COLUMN_NAME
+from h3ronpy.arrow import _to_arrow_array, _to_uint64_array
 from h3ronpy.h3ronpyrs import raster
-
-from .. import DEFAULT_CELL_COLUMN_NAME
-from . import _to_arrow_array, _to_uint64_array
-from .vector import cells_bounds, cells_to_wkb_polygons
+from h3ronpy.vector import cells_bounds, cells_to_wkb_polygons
 
 try:
     # affine library is used by rasterio
