+++ h3ronpy/python/h3ronpy/vector.py
@@ -2,10 +2,9 @@ from typing import Optional, Tuple
 
 from arro3.core import Array, DataType, RecordBatch
 
+from h3ronpy import ContainmentMode
+from h3ronpy.arrow import _to_arrow_array, _to_uint64_array
 from h3ronpy.h3ronpyrs import vector
-
-from .. import ContainmentMode
-from . import _to_arrow_array, _to_uint64_array
 
 
 def cells_to_coordinates(arr, radians: bool = False) -> RecordBatch:
